requirements.txt
# Core ML/AI
transformers>=4.35.0
torch>=2.0.0
onnxruntime>=1.16.0

# Vector Search
faiss-cpu>=1.7.4

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
slowapi>=0.1.9

# OpenAI
openai>=1.3.0

# AWS
boto3>=1.33.0

# Image Processing
Pillow>=10.1.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utilities
numpy>=1.24.0
aiofiles>=23.2.1

# UI
streamlit>=1.28.0

# Development
pytest>=7.4.0
httpx>=0.25.0
tqdm>=4.66.0